requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.16
einops>=0.7.0
opencv-python>=4.8.1.78
scikit-learn>=1.3.2
matplotlib>=3.8.2
transformers>=4.36.2
accelerate>=0.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
torchinfo>=1.5.0
pyyaml>=6.0
tqdm>=4.67.1
numpy>=1.24.3
Pillow>=10.1.0
